tweak to MOOSPlayBackV2.cpp
git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@10624 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ MOOS_Jul2724/MOOSToolsUI/Tools/Graphical/uPlayBack/MOOSPlayBackV2.cpp
@@ -285,7 +285,8 @@ bool CMOOSPlayBackV2::MessageFromLine(const std::string & sLine, CMOOSMsg &Msg)
         if(sData.find("<MOOS_BINARY>") !=std::string::npos)
         {
             //Msg.MarkAsBinary();
-            long long nOffset;
+            //long long nOffset;
+            double nOffset;
             if(!MOOSValFromString(nOffset,sData,"Offset"))
                 return MOOSFail("badly formed MOOS_BINARY indicator - missing \"Offset=xyz\"");
